Refact: useDragAndDrop hook
