+++ src/components/Selector.js
@@ -2,43 +2,17 @@ import Input from './Input';
 import ListItem from './ListItem';
 import StackedList from './StackedList';
 import Title from './Title';
-import { useRef, useState } from 'react';
+import { useState } from 'react';
+import useDragAndDrop from '../hooks/useDragAndDrop';
 
 export default function Selector({ list, isDraggable, order }) {
   const [items, setItems] = useState(list);
-
-  const draggingItem = useRef();
-  const dragOverItem = useRef();
-
-  const handleDragStart = (e, position) => {
-    draggingItem.current = position;
-    isDraggable[order] = false;
-  };
-
-  const onDragEnter = (e, position) => {
-    if (!(order === '1' ? isDraggable[0] : isDraggable[1])) {
-      return;
-    } else {
-      handleDragEnter(e, position);
-    }
-  };
-
-  const handleDragEnter = (e, position) => {
-    dragOverItem.current = position;
-
-    const itemsCopy = [...items];
-    const draggingItemContent = itemsCopy[dragOverItem.current];
-    itemsCopy.splice(dragOverItem.current, 1);
-    itemsCopy.splice(draggingItem.current, 0, draggingItemContent);
-    draggingItem.current = dragOverItem.current;
-    dragOverItem.current = null;
-
-    setItems(itemsCopy);
-  };
-
-  const handleDrop = () => {
-    isDraggable[order] = true;
-  };
+  const { handleDragStart, onDragEnter, handleDrop } = useDragAndDrop({
+    isDraggable,
+    order,
+    setItems,
+    items,
+  });
 
   return (
     <div
